organize into components folder
