+++ packages/ui/tsup.config.ts
@@ -48,8 +48,8 @@ export default defineConfig(overrideOptions => {
       'primitives/connection': 'src/primitives/connection.tsx',
       'primitives/field': 'src/primitives/field.tsx',
       'primitives/seperator': 'src/primitives/seperator.tsx',
-      'sign-in/index': 'src/sign-in/index.tsx',
-      'sign-up/index': 'src/sign-up/index.tsx',
+      'components/sign-in/index': 'src/components/sign-in/index.tsx',
+      'components/sign-up/index': 'src/components/sign-up/index.tsx',
     },
     external: ['react', 'react-dom'],
     format: ['cjs', 'esm'],
